+++ prover/src/common/prover/evm.rs
@@ -32,7 +32,7 @@ impl Prover {
                 )
                 .map_err(|err| anyhow!("Failed to construct compression circuit: {err:?}"))?;
 
-                let result = self.gen_evm_proof(id, degree, &mut rng, circuit);
+                let result = self.gen_evm_proof(id, degree, &mut rng, circuit, output_dir);
 
                 if let (Some(output_dir), Ok(proof)) = (output_dir, &result) {
                     proof.dump(output_dir, &name)?;
@@ -49,6 +49,7 @@ impl Prover {
         degree: u32,
         rng: &mut (impl Rng + Send),
         circuit: C,
+        output_dir: Option<&str>,
     ) -> Result<EvmProof> {
         Self::assert_if_mock_prover(id, degree, &circuit);
 
@@ -60,7 +61,10 @@ impl Prover {
         let instances = circuit.instances();
         let num_instance = circuit.num_instance();
         let proof = gen_evm_proof_shplonk(params, pk, circuit, instances.clone(), rng);
+        let evm_proof = EvmProof::new(proof, &instances, num_instance, Some(pk))?;
 
-        EvmProof::new(proof, &instances, num_instance, Some(pk))
+        crate::evm::gen_evm_verifier::<C>(params, pk.get_vk(), &evm_proof, output_dir);
+
+        Ok(evm_proof)
     }
 }
